backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
aioredis==2.0.1

# AI/ML - Speech Recognition
openai-whisper==20231117
# faster-whisper==0.10.0  # Requires av which needs C++ build tools

# AI/ML - NLP
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2

# AI/ML - Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
numpy==1.26.3
scipy==1.12.0

# AI/ML - Classification
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Language Support
indic-transliteration==2.3.44
langdetect==1.0.9
